backend/ai/requirements.txt
Flask
reportlab
google-generativeai
